requirements.txt
requests>=2.31.0
OTXv2>=1.5.12
python-whois>=0.8.0
ipwhois>=1.2.0
python-dotenv>=1.0.1

